WIP: fix multiple pages output
Current implementation is based on images processing and not text processing, resulting in large output files and sliced lines in bottom of the page. This commit is a WIP to fix this issue. Future implementation will be based on text processing, which should result in smaller output files and no sliced lines.